+++ docs/conf.py
@@ -68,7 +68,7 @@ author = 'Michael Ortner <magpylib@gmail.com>'
 # The short X.Y version
 version = ''
 # The full version, including alpha/beta/rc tags
-release = '1.2.0-beta'
+release = '1.2.1-beta'
 
 
 # -- General configuration ---------------------------------------------------